Rename CreateNotificationRequest to fit convention
