+++ src/main/java/teammates/ui/request/NotificationCreateRequest.java
@@ -1,0 +1,7 @@
+package teammates.ui.request;
+
+/**
+ * The request for creating a notification.
+ */
+public class NotificationCreateRequest extends NotificationBasicRequest {
+}
